Improve KeyboardInterrupt handling in testing framework
Testing should be aborted.  But in practice (with recent changes) only
the current top-level test group was aborted.

Also, the "Testing aborted." error message was logged immediately, and
then followed by various clean-up logging, and the test summary (which
included the "error" that testing was aborted in the "Errors logged"
counter.)  Re-arrange so that there is no test summary, and so that the
"Testing aborted." error message is the last thing logged, for clarity.