Don't loose ServerAdmin field in site config on upgrade
Before this fix the upgrade script didn't load any email at all
so data["installation.admin.email"] ended up as None when the
template for the apache2 site configuration file was populated.
This triggered a spurious diff/install query (even if nothing
related to the site config had changed) and the ServerAdmin would
be lost.

The apache2 ServerAdmin must have a single email only, so we
pick the email from the _first_ system recipient. And if that
has the from "Fullname <email>" we select only the actual email.

The regression being fixed by this commit was introduced by:
32f27f5 ("Separate the initial admin account from system message delivery")

Fixes: https://bugs.critic-review.org/show_bug.cgi?id=15