Add support for FilterHook extension roles
A FilterHook extension role allows an extension to implement additional
filter types (in addition to the native filter types "reviewer", "watcher"
and "ignored".)

When a user has installed the extension and has added such a filter, and
the filter matches changes in commits added to a review (including when
the review is first created) the extension role's function is called
(asynchronously) with the arguments "filter data", the Review object, a
CommitSet representing the added commits, and a list of File objects
representing the set of matched files that were touched by those commits.

The "filter data" is a string value that the user sets as part of the
filter (in the "Add filter" dialog) if the extension has requested it.
The extension requests it by defining the "DataDescription" setting in the
role's section in the MANIFEST file.  (If no such description is given,
the user is not presented with an input field for the "filter data" value,
and the value passed to the role's function is always null.)