Add basic test for tracking reviews
Tests creation of a tracking review, and a straight-forward rebase of it,
and also tests disabling and enabling of the tracking, and checks that the
tracking (and its status) is correctly reflected on the review front-page.

To make disabling/enabling easier, have the /submitreview operation return
the tracked branch id when creating a tracking review.