Make temporary references truly temporary
Some operations create temporary references (refs/temporary/$SHA1 or
refs/commit/$SHA1 traditionally,) for instance to be able to fetch a
specific commit into a temporary cloned repository.  These temporary
references were never cleaned up, making them more permanent than
temporary in practice.

This commit introduces the utility function Repository.temporaryref()
with acts as a context manager and returns a temporary reference name
that is deleted when the context manager's __exit__() is called.  Code
that used to create "temporary" references is modified to use this
instead.

Also, the utility function Repository.fetchTemporaryFromRemote() is
adjusted to also act as a context manager that returns the fetched SHA-1,
and also makes sure to clean up the temporary reference name it creates
in the context manager's __exit__().

The /fetchremotebranch operation, which is the only current user of
Repository.fetchTemporaryFromRemote(), is adjusted to add a permanent
keepalive reference to the fetched commits instead.  In practice, the
operation returns SHA-1s back to web front-end clients, that will later
supply them as arguments to /submitreview, so the commits need to be
referenced at least semi-persistently.  We don't have semi-persistently,
so make it fully persistently instead.  (It's likely the commits will be
used to create a review, in which case they can never be GC:ed anyway.)

Also add a migration script that deletes all existing refs/temporary/
and refs/commit/ references in every repository.