Pack keepalive refs into a single ref
Before running the nightly maintenance GC on each repository, pack all
"loose" keepalive refs (refs/keepalive/$SHA1) into a single keepalive
ref (refs/keepalive-chain), and delete all the loose keepalive refs.
This is done to reduce the total number of refs in the repository,
which improves the performance of various operations on it, and also
reduces network traffic when pushing to the repository.

The single keepalive ref will always point to a chain of "manufactured"
merge commits, that point to the empty tree, and whose parents will be
another such merge commit (or a root commit) and the commit we wish to
keep alive.

Also add a migration script that converts all existing refs/replays/*
refs into refs/keepalive/* commits, so that they are then packed and
deleted by this packing mechanism too.