Fix /rebasetrackingreview output for rebases that introduce changes
Due to a typo/copy-paste-error the "Has unexpected changes!" status
would never be reported; that case would say "Clean." instead.  Also,
there would be no link to a diff displaying the changes.

The outcome of the rebase was not affected, but this might cause users
to not notice invalid rebases before-hand when they could have if the
system had helped them out a little more.