Stop changing ServerAdmin value on first upgrade
When installing, data["installation.admin.email"] was set to the admin
user's (initial) email address, as stored in the users table.  But on
upgrade, it was set to configuration.base.SYSTEM_RECIPIENTS[0]'s email
address, which is separate (although the same post-install, by default.)

To avoid the resulting spurious site file change on first upgrade, that
happens unconditionally if admin email and system recipient are not the
same, do not touch data["installation.admin.email"] when upgrading, and
reuse the value as stored in .install.data when installing instead.