Add generic support for "clickable" table cells
When a table cell contains a link, it sometimes makes sense to redirect
clicks on the table cell (but not on the link) to the link.

This was done on the commit summary cell in branch logs, using custom
code.  Extract that code and make it available for re-use, by setting
the "clickable" class on table cells.