Run "late upgrade" tests whenever migration scripts are added/modified
Typically, the point of late upgrading is upgrading with some data in the
database or branches in the Git repository, and typically the only part
of upgrading that cares about such things are migration scripts.

And, when there are migration scripts, they almost always do care, so it
makes sense to always run the late upgrading tests then.