Restructure session handling in testing.frontend.Frontend
Previously, Frontend.signin() was a bit confusing, you could optionally
call it with username=None, in which case it didn't sign in.  This was
used when the caller had manually triggered an operation (or loaded a
page) that returned a session id cookie, and just wanted a context object
that signed out automatically.  Additionally, signin() never checked that
a session cookie was actually received.

The new design is:

 - Each call to Frontend.page()/Frontend.operation() will set a new
   attribute, Frontend.pending_session_id to the value of the session id
   cookie received, or None.

 - A new method, Frontend.session(), asserts that there is a pending
   session id, pushes it onto a stack of current session ids and returns
   a context object that signs out and pops the session id from the stack
   on __exit__.

 - Another new method, Frontend.no_session(), returns a context object
   that disables the current session when entered and restores it when
   exited.

 - The simplified Frontend.signin() (which no longer supports being called
   with username=None) simply executes the /validatelogin operation and
   defers to Frontend.session() to do the rest.

 - Frontend.signout() is inlined into Frontend.session() and dropped.

The only test that is significantly affected by this is 003-oauth.py,
which was the user of Frontend.signin(username=None).  It's simplified as
a result.

The advantages of this redesign are that it becomes impossible to
accidentally have a session id stored and used for future requests, since
session id cookies aren't used automatically when received; and that it
becomes possible to have nested sessions (at least for different users.)