Internal: Support optional "prioritized" operation parameters
They won't be prioritized then, just optional.  Typical use-case of a
prioritized parameters, for instance "repository", is to make the
interpretation of other parameters, like commit references, meaningful.

If a repository parameter is optional, it probably isn't meant to fill
that kind of purpose, which is fine; just ignore that it's normally
prioritized.