Internal: Use User/Repository type checkers for /savesettings
This makes the operation somewhat more convenient to use from testing,
and is also the "modern" way of handling operation parameters.  To do
this, we needed to stop abusing the |user_id| parameter to double as
"set system defaults" when it has the magic (invalid) value -1.  Use an
explicit |defaults| parameter for that instead.