Fix race condition in 030-trackingreview.py
The test creates a review that tracks a remote branch using the
/submitreview operation, and then immediately rebases it to track a
different remote branch using the /rebasereview operation.

Since /submitreview also schedules an immediate update of the tracked
branch (by the branch tracker background service,) we need to
synchronize to avoid the branch tracker's update happening at the
wrong time, conflicting with the review rebase.