Fix "Work log" batch summary for reviewed/unreviewed lines
Instead of saying "reviewed 10 lines" when the batch actually marked 10
lines worth of changes as reviewed, we said so when the batch changed
the reviewed state of 10 lines worth of changes that are currently
reviewed.  In other words, any set of batches that changed the reviewed
state of the same changes would always say the same thing, "reviewed
N lines" or "unreviewed N lines", depending on the current state of the
changes.

Fix this by looking at the reviewfilechanges.to(_state) column instead
of the reviewfiles.state column when categorizing work done by a batch.