Remove bzip2 compression of database dump
When we make a backup database dump as part of upgrade (default when the
database schema will be updated) we compress the dump using bzip2.  This
is pretty silly, since we tell pg_dump to use its "custom" format, which
is already compressed.  The additional compression achieved by bzip2 is
marginal at best, and mostly just a waste of CPU time.