Add quickstart.py script
Allows a regular (non-privileged) user to start a Critic instance.
Uses the built-in Python modules wsgiref.simple_server instead of
Apache/mod_wsgi and sqlite3 instead of PostgreSQL.  Still depends on
the non-standard Python modules requests, passlib and pygments
installed, but those could presumably be installed using "pip install"
in a virtualenv, and only requests is a hard requirement.

Also add a testing mode (python -m testing --quickstart) where tests
are run against a quick-started instance of Critic instead of a full
instance installed in a VirtualBox VM.  This mode doesn't support all
tests, but large parts of the test suite can be run this way.