Fetch remote branches without a relay repository
Modify gitutils.Repository.fetchTemporaryFromRemote() to run directly in
the target repository, by fetching to a temporary ref (whose name includes
a timestamp and a some randomness) and processing the fetched commit by
calling index.processCommits().

Creating a relay copy is usually very cheap, but does become expensive if
the main repository has very many loose objects.  A single fetch directly
in the target repository should also be quite a bit faster than a clone,
a fetch in the clone, and then a push back to the target repository, even
in the best of cases.