Add --upgrade-after argument and drop 002-late-upgrade test group
Late upgrading is sometimes absolutely necessary testing, but the model
implemented by the 002-late-upgrade group is inherently inflexible and
impractical.  What tests need to run before the upgrade, and what to test
after the upgrade, depends entirely on the upgrade and the particular
migration scripts it will run, and isn't practical to "hard-code" with a
test group like 002-late-upgrade.  Also, an extra group adds significant
cost to testing, often to little or no use.

The new --upgrade-after means any set of tests can run, and the upgrade
can be "inserted" anywhere in that set of tests.  For now, this will only
be useful for manual testing, but it would be possible to add some
mechanism by which a commit (via some checked in control file) can declare
useful test runs, that the automatic testing system could use to run
additional custom tests when testing that particular commit.