Clean up handling of equivalent merges and replayed rebases
When performing a move rebase, we will either create an "equivalent merge"
(for a fast-forward rebase) or a "replayed rebase" (otherwise.)  In the
former case, we'd update the review rebase's "old head" to point to the
equivalent merge, as if it had been pushed to the review before the rebase.
A replayed rebase wasn't directly recorded in the database at all, instead
its existence was deduced from there being a changeset added to the review
of type "conflicts" whose child commit was the new head of the rebase.

This was generally messy, and error-prone to work with.

Instead of all this, add two explicit columns to the reviewrebases table,
equivalent_merge and replayed_rebase, and set the appropriate one to point
to the generated commit when performing a rebase.  Adjust code elsewhere
to use the values of these columns.