Mark comment read when specifying silent operation for Batch.finish()
When completing a Batch in a 'silent' mode (FinishData.silent == true),
in addition to not sending mails, also don't make added comments show
up as unread. Reasoning is that reviews with unread comments are
specially marked on the Dashboard page and if the changes were not
important enough to bother user with emails then he should also not
be bothered in there.