Make review branch archival more robust
It used to crash if it encountered a review branch that didn't actcually
exist in the repository, and if the (missing) review branch according to
the database referenced a commit that didn't exist in the repository, and
if review branch in the repository pointed to a different commit than it
was supposed to according to the database.

Handle the first two classes of errors by simply ignoring them, and just
setting the branch's "is archived" flag to true.

Handle the last class of errors by silently not archiving the branch (or
setting its "is archived" flag.)  If called because an archival was
scheduled, the branch would not be scheduled for archival again (by this)
so we'd just leave it as-is until something else happens.