Improve quickstart system restart mechanism for testing
A quickstarted system by default continuously checks for changed files and
restarting the system automatically to "update."  During such a restart,
there's a brief moment when the system is unavailable, that tends to break
testing.

So instead of doing that (while testing,) restart only when SIGUSR1 is
received.

Also fix the testing framework's --pause-upgrade-retry mode for use with
--quickstart, by sending SIGUSR1 to the instance process and waiting for
it to respond.  Previously, --pause-upgrade-retry only worked with full
VirtualBox instances, for no good reason.