Improve handling of unexpected exceptions raised by individual tests
While writing test cases, they are often broken in such a way that they
raise "unexpected" exceptions (not sub-classes of testing.TestFailure.)
Such errors were treated as very "fatal", completely aborting the testing
and overriding for instance --pause-upgrade-retry.

It is unlikely, in fact, that such an exception is that fatal; much more
likely it's just a coding error in the test that doesn't affect whether
the test can be retried, and even less whether other tests can be run.

Thus, instead of treating unexpected exceptions as extra fatal, treat them
as testing.TestFailure exceptions and just log the stack trace as an error.