Add API modules api.comment and api.reply
The api.comment.Comment class represents review comments (issues and notes)
and the api.reply.Reply class represents replies to those comments.

Note: This structure and naming is different from the old/internal variant
where comments are named "comment chains", replies are named "comments" and
where the initial comment text is represented as a reply (sort of.)  The
new variant is simpler and better.  There's no plan to adjust the internal
code or database schema to reflect this change.

Also note: "Code comments" are not fully implemented since the API doesn't
have representation of changesets yet.  In practice, this means that the
Comment.location attribute raises on reading for code comments.  Also the
api.comment.fetchAll() function raises if the filtering arguments 'commit'
or 'changeset' are used.