Fix infinite loop in api.commitset.CommitSet.date_ordered
If the sorting function encountered the sequence of commits A-B-X where
A is the parent of B is the parent of X, and X's commit timestamp is
earlier than both A and B, it would get stuck forever alternating between
trying to emit A and B (only) failing A since B hasn't been emitted and
failing B since X hasn't been emitted.

Also extend the unit test slightly to trigger this situation.