Restructure handling of context objects in JSON API implementation
Instead of passing the result from the previous path segment on to the
resource class handling the next path segment, collect a mapping
{ resource name => context object } of all previous path segments.

This mapping can further be updated by a resource class that implicitly
references other resources.  A review resource, for instance, now adds its
repository and branch to the context.  Having the branch in the mapping in
turn made it trivial to support reviews/N/commits by simply reusing the
resource class that handles branches/N/commits.