JSON API: Expose review log partitions and review rebases
Rebases have unique ids and are exposed as primary resources, via the path
/api/v1/rebases or /api/v1/reviews/N/rebases.

Partitions are exposed as secondary resources in the review resource, as
a list assigned to the 'partitions' property.  They in turn reference the
rebases.

Additionally, the 'commits' property in the review resource is dropped.
The exact same set of commits is now listed in the partitions list, just
split into different sub-lists (if the review has been rebased.)  The
current set of commits on the review branch (which the 'commits' property
did not represent) is accessible via /api/v1/reviews/N/commits.