Restructure handling of linked resources in JSON API implementation
Internally, the default way of recording a linked resource was by its
resource name and the resource id.  Recording the resource name and the
actual internal value was also supported, however.

In practice, when recording a linked resource, the resource id was always
fetched from the internal value, so recording the internal value was
actually less code.  And recording by id meant that for some resources
(commits, at least) it was actually possible to record the same resource
twice, once by id and once by value.

So, instead, always record linked resources by value.  This is more
straight-forward for the recording code, and also simpler when processing
linked resources, since there's no need to resolve the id back to a
resource value.  In some cases (commits, again) this resolving is somewhat
tricky, since going from commit id to commit requires a repository.  In
some cases, like when listing multiple reviews (potentially in different
repositories) and their linked commits, this would have been impossible to
handle by having a (single) repository in the context.