Fix comment propagation past fast-forward move rebases
Adjust ReviewRebases.fromOldHead() to handle the case that the "old head"
is an equivalent merge commit.  It used to, since equivalent merge commits
used to actually be the old heads of their rebases.  When that was changed,
comment propagation, the only user of fromOldHead(), was inadvertently
affected (broken, in fact.)

This fixes a regression introduced by

  0103797d "Clean up handling of equivalent merges and replayed rebases"