Restructure prerequisite installation
The main difference is that prerequisites are now installed after the
"prepare" step (by the 'prereqs' module's install()/upgrade() functions.)
This makes sense, since what prerequisites we install depend on choices
made in other modules' prepare() (notably the 'config' module's.)

Prior to this, the choice is first made whether to install an optional
prerequisite, and then later the choice is made whether to use it
(although that choice is only available if it was installed.)  With this
patch, the main choice is made first, and then the affected prerequisite
is either not mentioned at all, or requested/installed unconditionally.

The code is also refactored to be a bit nicer and less repetitive.