Make database cursor handling a bit more flexible
It now becomes possible to create an updating cursor while another is in
scope, but only if the nested updating cursor is for a subset of the tables
updated by the outer updating cursor. The nested updating cursor does not
commit the transaction.

It also becomes possible to create an "old-style" cursor while an updating
cursor is in scope. The returned cursor is restricted such that it can
only modify tables updated by the current updating cursor.