Add Session.refresh() / Database.refresh() methods
Session.refresh() clears the dbutils.User object cache, since it might be
stale.  Database.refresh() in addition commits the database transaction
(unless that would be bad, in which case it raises an exception.)

The use-case is that another database transaction has been committed
(typically by another (background) process), and the current connection's
transaction needs to be reset to see the changes committed.

It's always safe to use in so far that it raises if it would actually
commit any changes to the database.