Share code for executing command during testing
Instance.execute() (virtualbox.py) was quite advanced, incrementally
logging output (stdout and stderr separately), making it possible to see
in the log when during the execution of a slow command different lines of
output was emitted.

It makes sense to handle execution of Git commands the same way since
they too take a long time sometimes, and the output received is of some
interest as well.