Extend API with a section for changing the system state
The API consists of the object api.transaction.Transaction, which is a
context manager object that performs queued changes and commits the
database transaction upon exit.

Currently, the transaction object has one public method, modifyUser(),
which returns an api.transaction.user.ModifyUser object, which has methods
for queuing changes to the user.  This currently consists of the methods
setFullname(), createFilter() and modifyFilter().  The latter in turn
returns an api.transaction.filters.ModifyRepositoryFilter object, with
methods for queuing changes to filters (including deleting the filter.)

The set of operations supported here is highly arbitrary; this is just a
shell that will be expanded over time to support more things as needed.