Add support for POST, PUT and DELETE in the JSON API
POST creates new resources.  The path should not identify a specific
resource, but can identify a list sub-resource.  The created resource
is returned (as if a GET request had been used to access it.)

PUT updates existing resources.  The path should identify a specific
resource to update.  Typically, the updated resource is returned.

DELETE deletes existing resources, or items in list sub-resources.  The
response is either a "204 No Content" or the modified resource (if an
item in a list sub-resource was deleted.)