Return "303 See Other" from /redirect
The /redirect URL is primarily used as the action of the login form, to
trigger the password saving mechanism.  (If not for that, we wouldn't
submit the form at all, since the actual login is done using XHR.)

To avoid issues with redirecting non-GET requests, the redirect was made
client-side by returning a "200 OK" response with a

  <meta http-equiv='refresh' content='0; $target'>

content.  But this also caused some issues.

Using a "303 See Other" response should be fine; it explicitly means "go
here, but use a GET request" (like "302 Moved Temporarily" used to, but
which was just never actually implementated that way.)