Add basic extensions support to API and JSON API
Can be used to list available extensions, optionally filtered by publisher
and/or user that has the extensions installed, or be accessed by id or key.

Only exposes very basic information about the extensions, and is mostly
added so that other parts of the APIs can refer to specific extensions.