Add support for access control profiles
This can be used to restrict how a single user, or anonymous users, can
access the system at all, but first of all it makes it possible to
restrict how the system can be accessed when authenticating with an access
token, which greatly improves the usefulness of access tokens.

For now, the only way of configuring an access control profile is using
the API / JSON API, and it only supports configuring the access control
profile that belongs to an access token.