Add DEFAULT NOW() to column comments.time
This makes sense since most other timestamp columns have a default,
and in particular all in dbschema.comments.sql.  It also fixes a
crash in the extension library code, which inserts rows into the
comments table without specifying a value for the time column, thus
violating its NOT NULL constraint.