Fix PASSWORD_HASH_SCHEMES in quick-started instance config
The value was a string (single hash scheme, same as default) instead
of a list of strings as intended.  This resulted in the creating a
CryptContext where the default hash scheme was deprecated (we
deprecate all schemes but the default) which makes no sense, and is
rejected by newer passlib versions.