Move password changing to authentication database hierarchy
This is where it belongs, since it's very much specific to the database
used. In particular, currently, password changing is only actually
supported when the "internal" database is used. (For LDAP, it could be
supported, but isn't.)

Also, make sure not to allow password changes if the user authenticated
using an access token.

Finally, correctly determine whether to display the "change password" UI
on /home. It was displayed whenever Critic did authentication itself
(rather than the host web server) but this was wrong e.g. when LDAP was
used.