Improve "create tracking review"'s "wrong remote" check
To avoid mistakes that risk adding significant bloat to repositories, we
don't allow fetching a branch from a remote that branches in another
repository already tracks, since most likely, that remote "belongs" to
the other repository, and either the local repository or the remote is
wrong. (IOW, the user meant to do something else.)

However, if branches in two different repositories already tracked
branches from a shared remote, then the remote would be rejected for
reviews in both repositories, which is hardly useful.

To avoid this, loosing the check somewhat to always allow a remote that
there are already branches in the selected local repository that tracks,
even if there are also branches in other repositories that track branches
from that remote.