Implement repository access check for extensions
Without this, a user could access any repository using the extension API
when executing an extension.

Since extensions can also typically access the entire file system, and
run arbitrary child processes, this is far from a complete fix, but it's
something. It does protect against a "malicious" user executing a "good"
extension, but it doesn't protect against a "malicious" extension.